Move cart related views to payments app
