+++ backend/events/urls.py
@@ -23,7 +23,6 @@ router.register('ticket', views.TicketViewSet, basename="ticket")
 urlpatterns = router.urls
 
 urlpatterns += [
-    path('cart', views.ShoppingCartView.as_view(), name='cart'),
     path(
         'event/<int:pk>/send-notification',
         views.CreateNotificationView.as_view(),
