+++ backend/events/views.py
@@ -1,11 +1,8 @@
-from django.core.cache import cache
 from django.utils import timezone
 
-from rest_framework import exceptions, mixins, permissions, status, viewsets
+from rest_framework import exceptions, mixins, permissions, viewsets
 from rest_framework.generics import CreateAPIView
 from rest_framework.filters import SearchFilter
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from events import permissions as event_permissions
 from events import models, serializers
@@ -148,47 +145,6 @@ class TicketViewSet(viewsets.ModelViewSet):
             raise exceptions.PermissionDenied(msg)
 
         return super().perform_create(serializer)
-
-
-class ShoppingCartView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
-
-    def get_cart(self, id):
-        data = cache.get(id)
-        if not data:
-            return Response({'error': 'No data assigned to the given key.'},
-                            status=status.HTTP_400_BAD_REQUEST)
-        return data
-
-    def get(self, request, *args, **kwargs):
-        data = self.get_cart(request.user.id)
-        if isinstance(data, Response):
-            return data
-
-        serializer = serializers.CartSerializer(data=data)
-        serializer.is_valid(raise_exception=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def post(self, request, *args, **kwargs):
-        serializer = serializers.CartSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        cache.set(request.user.id, serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    def put(self, request, *args, **kwargs):
-        serializer = serializers.CartSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        old_data = cache.get(request.user.id)
-        cache.set(request.user.id, serializer.data)
-
-        if old_data:
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    def delete(self, request, *args, **kwargs):
-        cache.delete(request.user.id)
-        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class CreateNotificationView(CreateAPIView):
